v73

+++ hlt/scorer.hpp
@@ -19,7 +19,7 @@ namespace hlt
 	public:
 		vector<vector<double>> grid_score_highway;
 		vector<vector<int>> grid_score_move;
-		vector<vector<int>> grid_score_enemies;
+		vector<vector<double>> grid_score_enemies;
 		vector<vector<double>> grid_score_extract;
 		vector<vector<double>> grid_score_dropoff;
 		vector<vector<double>> grid_score_extract_smooth;
@@ -39,7 +39,7 @@ namespace hlt
 		Scorer(int height, int width) : halite_initial(0), halite_total(0), halite_percentile(0) 
 		{
 			grid_score_move = vector<vector<int>>(height, vector<int>(width, 0));
-			grid_score_enemies = vector<vector<int>>(height, vector<int>(width, 0));
+			grid_score_enemies = vector<vector<double>>(height, vector<double>(width, 0.0));
 			grid_score_ships_nearby = unordered_map<PlayerId, vector<vector<double>>>();
 
 			grid_score_extract = vector<vector<double>>(height, vector<double>(width, 0.0));
@@ -68,7 +68,7 @@ namespace hlt
 		void add_self_ships_to_grid_score(shared_ptr<Ship> ship, const Position& position);
 		void flush_grid_score(const Position& position) { grid_score_move[position.y][position.x] = 0; }
 		int get_grid_score_move(const Position& position) const { return grid_score_move[position.y][position.x]; }
-		int get_grid_score_enemies(const Position& position) const { return grid_score_enemies[position.y][position.x]; }
+		double get_grid_score_enemies(const Position& position) const { return grid_score_enemies[position.y][position.x]; }
 		void update_grid_score_move(const Game& game);
 		void update_grid_score_enemies(const Game& game);
